chore: header cleanup and organisation
