+++ pages/results.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components'
 import { RootState } from '../redux/types'
 
 import { Main } from '../components/layout/Main'
-import { SlimHeader } from '../components/header/slim/SlimHeader'
+import { Header } from '../components/header/Header'
 import { Heading } from '../components/heading/Heading'
 import NoResults from '../components/no-results/NoResults'
 
@@ -160,7 +160,7 @@ const ResultsPage = () => {
 
     return (
         <Main page={"results"}>  
-            <SlimHeader id="home" logo="logo-icon-white.svg" acronym="logo-acronym-white.svg" />
+            <Header />
 
             {!budget && (
                 <NoResults cause={'budget'} />
